shop admin edit, search and save working

+++ admin/src/main/java/com/example/admin/data/repo/AdminRepository.kt
@@ -27,7 +27,7 @@ interface AdminRepository {
 
     suspend fun generateNewId(
         location: String
-    ): Int?
+    ): Long?
 
     suspend fun searchTerm(
         collection: String,
